+++ src/utils.ts
@@ -1,11 +1,9 @@
-import { unmarshalPublicKey } from '@libp2p/crypto/keys'
+import { publicKeyFromProtobuf, publicKeyToProtobuf } from '@libp2p/crypto/keys'
+import { UnexpectedPeerError, type PrivateKey, type PublicKey } from '@libp2p/interface'
 import { type Uint8ArrayList } from 'uint8arraylist'
-import { equals, toString } from 'uint8arrays'
 import { concat as uint8ArrayConcat } from 'uint8arrays/concat'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
-import { UnexpectedPeerError } from './errors.js'
 import { type NoiseExtensions, NoiseHandshakePayload } from './proto/payload.js'
-import type { PrivateKey } from '@libp2p/interface'
 
 export async function createHandshakePayload (
   privateKey: PrivateKey,
@@ -15,7 +13,7 @@ export async function createHandshakePayload (
   const identitySig = await privateKey.sign(getSignaturePayload(staticPublicKey))
 
   return NoiseHandshakePayload.encode({
-    identityKey: privateKey.public.bytes,
+    identityKey: publicKeyToProtobuf(privateKey.publicKey),
     identitySig,
     extensions
   })
@@ -24,15 +22,14 @@ export async function createHandshakePayload (
 export async function decodeHandshakePayload (
   payloadBytes: Uint8Array | Uint8ArrayList,
   remoteStaticKey?: Uint8Array | Uint8ArrayList,
-  remoteIdentityKey?: Uint8Array | Uint8ArrayList
+  remoteIdentityKey?: PublicKey
 ): Promise<NoiseHandshakePayload> {
   try {
     const payload = NoiseHandshakePayload.decode(payloadBytes)
-    if (remoteIdentityKey) {
-      const remoteIdentityKeyBytes = remoteIdentityKey.subarray()
-      if (!equals(remoteIdentityKeyBytes, payload.identityKey)) {
-        throw new Error(`Payload identity key ${toString(payload.identityKey, 'hex')} does not match expected remote identity key ${toString(remoteIdentityKeyBytes, 'hex')}`)
-      }
+    const publicKey = publicKeyFromProtobuf(payload.identityKey)
+
+    if (remoteIdentityKey?.equals(publicKey) === false) {
+      throw new Error(`Payload identity key ${publicKey} does not match expected remote identity key ${remoteIdentityKey}`)
     }
 
     if (!remoteStaticKey) {
@@ -40,7 +37,6 @@ export async function decodeHandshakePayload (
     }
 
     const signaturePayload = getSignaturePayload(remoteStaticKey)
-    const publicKey = unmarshalPublicKey(payload.identityKey)
 
     if (!(await publicKey.verify(signaturePayload, payload.identitySig))) {
       throw new Error('Invalid payload signature')